requirements.txt
fastapi
uvicorn
fastembed
faiss-cpu
numpy
httpx
python-dotenv
pydantic
gunicorn
